Extracts the MessageBuilder
